Extract market activity record creation logic
